clean up APL check
